+++ database-ms/src/main/java/org/dice_research/sask/database_ms/DbController.java
@@ -25,9 +25,9 @@ import org.apache.jena.update.UpdateFactory;
 import org.apache.jena.update.UpdateProcessor;
 import org.apache.jena.update.UpdateRequest;
 import org.apache.log4j.Logger;
-import org.dice_research.sask.database_ms.RDFTriples.AutoIndexDTO;
-import org.dice_research.sask.database_ms.RDFTriples.EndPointParameters;
-import org.dice_research.sask.database_ms.RDFTriples.TripleDTO;
+import org.dice_research.sask.database_ms.rdftriples.AutoIndexDTO;
+import org.dice_research.sask.database_ms.rdftriples.EndPointParameters;
+import org.dice_research.sask.database_ms.rdftriples.TripleDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.http.HttpEntity;
