--- database-ms/src/main/java/org/dice_research/sask/database_ms/RDFTriples/TripleDTO.java
@@ -1,22 +1,0 @@
-/**
- * 
- *
- * @author Sepide Tari
- * @author Suganya Kannan
- *
- *
- **/
-package org.dice_research.sask.database_ms.RDFTriples;
-
-public class TripleDTO {
-	private String triple = "";
-
-	public String getTriple() {
-		return triple;
-	}
-
-	public void setTriple(String triple) {
-		this.triple = triple;
-	}
-
-}
