+++ database-ms/src/main/java/org/dice_research/sask/database_ms/rdftriples/TripleDTO.java
@@ -1,0 +1,22 @@
+/**
+ * 
+ *
+ * @author Sepide Tari
+ * @author Suganya Kannan
+ *
+ *
+ **/
+package org.dice_research.sask.database_ms.rdftriples;
+
+public class TripleDTO {
+	private String triple = "";
+
+	public String getTriple() {
+		return triple;
+	}
+
+	public void setTriple(String triple) {
+		this.triple = triple;
+	}
+
+}
